feat: Implemented 2 seperate classes to represent the items and the current transaction
Because:
-The code was starting to look a bit messy, this cleans it
up

There are now 2 classes, item and transaction. The item represents
an individual item which can be purchased. The transaction object stores
those items, as well as groups relevant calculation logic and values. It
may have been more logical to place the functions in the cash register
class, and just have the Transaction class store tbe relevant values.

Closes Allow multiple items in each transaction